requirements.txt
streamlit
openai
huggingface_hub
python-dotenv
requests